Main part of Kvintakord's rewrite
- Split into seperate classes
- Kvintakord is now implemented non-statically
- Overhauled Spotify implementation
