+++ src/main/java/com/pattexpattex/servergods2/commands/slash/fun/EmoteCmd.java
@@ -1,6 +1,6 @@
 package com.pattexpattex.servergods2.commands.slash.fun;
 
-import com.pattexpattex.servergods2.core.BotException;
+import com.pattexpattex.servergods2.core.exceptions.BotException;
 import com.pattexpattex.servergods2.core.commands.BotSlash;
 import com.pattexpattex.servergods2.util.BotEmoji;
 import com.pattexpattex.servergods2.util.FormatUtil;
