+++ src/main/java/com/pattexpattex/servergods2/commands/slash/fun/GiveawayCmd.java
@@ -1,7 +1,7 @@
 package com.pattexpattex.servergods2.commands.slash.fun;
 
 import com.pattexpattex.servergods2.core.Bot;
-import com.pattexpattex.servergods2.core.BotException;
+import com.pattexpattex.servergods2.core.exceptions.BotException;
 import com.pattexpattex.servergods2.core.commands.BotSlash;
 import com.pattexpattex.servergods2.core.giveaway.Giveaway;
 import com.pattexpattex.servergods2.core.giveaway.GiveawayManager;
