+++ src/main/java/com/pattexpattex/servergods2/core/config/GuildConfig.java
@@ -1,6 +1,6 @@
 package com.pattexpattex.servergods2.core.config;
 
-import com.pattexpattex.servergods2.core.Kvintakord;
+import com.pattexpattex.servergods2.core.kvintakord.Kvintakord;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Role;
 import net.dv8tion.jda.api.entities.TextChannel;
