+++ src/main/java/com/pattexpattex/servergods2/core/kvintakord/discord/AudioSendHandler.java
@@ -1,0 +1,41 @@
+package com.pattexpattex.servergods2.core.kvintakord.discord;
+
+import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
+import com.sedmelluq.discord.lavaplayer.track.playback.MutableAudioFrame;
+import org.jetbrains.annotations.Nullable;
+
+import java.nio.Buffer;
+import java.nio.ByteBuffer;
+
+public class AudioSendHandler implements net.dv8tion.jda.api.audio.AudioSendHandler {
+
+    private final AudioPlayer audioPlayer;
+    private final ByteBuffer buffer;
+    private final MutableAudioFrame frame;
+
+    public AudioSendHandler(AudioPlayer audioPlayer) {
+
+        this.audioPlayer = audioPlayer;
+        this.buffer = ByteBuffer.allocate(1024);
+        this.frame = new MutableAudioFrame();
+
+        this.frame.setBuffer(buffer);
+    }
+
+    @Override
+    public boolean canProvide() {
+        return audioPlayer.provide(frame);
+    }
+
+    @Nullable
+    @Override
+    public ByteBuffer provide20MsAudio() {
+        buffer.flip();
+        return buffer;
+    }
+
+    @Override
+    public boolean isOpus() {
+        return true;
+    }
+}
